+++ Models/Route.cs
@@ -7,7 +7,7 @@ namespace hajk.Models
         [PrimaryKey, AutoIncrement]
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Distance { get; set; }
+        public float Distance { get; set; }
         public int Ascent { get; set; }
         public string Description { get; set; }
         public string WayPoints { get; set; }
